add full work with upload

+++ index.js
@@ -1,3 +1,4 @@
+"use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
@@ -34,18 +35,21 @@ var __generator = (this && this.__generator) || function (thisArg, body) {
         if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
     }
 };
-var _this = this;
+exports.__esModule = true;
+var AWS = require('aws-sdk');
 var exec = require('child_process').exec;
 var moment = require('moment');
 var axios = require('axios');
 var ffmpegPath = require('@ffmpeg-installer/ffmpeg').path;
 var fs = require('fs');
-var path = require('path');
+require('dotenv').config();
 var backetDefaultBase = 'https://storage.yandexcloud.net';
+var backetDefaultName = 'videopreviewer';
 var OUT_FOLDER = 'tmp';
 var OUT_FORMAT = 'jpg';
-var getBlobLocalFile = function (path) { return __awaiter(_this, void 0, void 0, function () {
-    var data, e_1;
+var isCachingImages = false;
+var getBlobLocalFile = function (path) { return __awaiter(void 0, void 0, void 0, function () {
+    var e_1;
     return __generator(this, function (_a) {
         switch (_a.label) {
             case 0:
@@ -54,10 +58,8 @@ var getBlobLocalFile = function (path) { return __awaiter(_this, void 0, void 0,
                 _a.label = 1;
             case 1:
                 _a.trys.push([1, 3, , 4]);
-                return [4 /*yield*/, fs.promises.readFile(path)];
-            case 2:
-                data = _a.sent();
-                return [2 /*return*/, String(data)];
+                return [4 /*yield*/, fs.promises.readFile(path, 'binary')];
+            case 2: return [2 /*return*/, _a.sent()];
             case 3:
                 e_1 = _a.sent();
                 console.log('getBlobLocalFile', e_1);
@@ -66,7 +68,7 @@ var getBlobLocalFile = function (path) { return __awaiter(_this, void 0, void 0,
         }
     });
 }); };
-var getCacheImage = function (path) { return __awaiter(_this, void 0, void 0, function () {
+var getCacheImage = function (path) { return __awaiter(void 0, void 0, void 0, function () {
     var cachedImage, e_2;
     return __generator(this, function (_a) {
         switch (_a.label) {
@@ -81,13 +83,12 @@ var getCacheImage = function (path) { return __awaiter(_this, void 0, void 0, fu
                 return [2 /*return*/, { cachedImage: cachedImage, isCached: !!cachedImage }];
             case 3:
                 e_2 = _a.sent();
-                // console.log('getCacheImage', e)
                 return [2 /*return*/, { image: '', isCached: false }];
             case 4: return [2 /*return*/];
         }
     });
 }); };
-var removeFile = function (path) { return __awaiter(_this, void 0, void 0, function () {
+var removeFile = function (path) { return __awaiter(void 0, void 0, void 0, function () {
     var e_3;
     return __generator(this, function (_a) {
         switch (_a.label) {
@@ -109,7 +110,7 @@ var removeFile = function (path) { return __awaiter(_this, void 0, void 0, funct
 var getAndCacheImageByVideo = function (uri, fileNameWithoutWxt, options) {
     if (uri === void 0) { uri = ''; }
     if (fileNameWithoutWxt === void 0) { fileNameWithoutWxt = ''; }
-    return __awaiter(_this, void 0, void 0, function () {
+    return __awaiter(void 0, void 0, void 0, function () {
         var resultFileUrl, output_path, _a, cachedImage, isCached, _b, _c, width, _d, height, _e, seconds, time, args, isCreated, imageData;
         return __generator(this, function (_f) {
             switch (_f.label) {
@@ -120,12 +121,14 @@ var getAndCacheImageByVideo = function (uri, fileNameWithoutWxt, options) {
                     return [4 /*yield*/, getCacheImage(output_path)];
                 case 1:
                     _a = _f.sent(), cachedImage = _a.cachedImage, isCached = _a.isCached;
-                    if (!isCached) return [3 /*break*/, 3];
+                    if (!isCached) return [3 /*break*/, 4];
+                    if (!!isCachingImages) return [3 /*break*/, 3];
                     return [4 /*yield*/, removeFile(output_path)];
                 case 2:
                     _f.sent();
-                    return [2 /*return*/, cachedImage];
-                case 3:
+                    _f.label = 3;
+                case 3: return [2 /*return*/, cachedImage];
+                case 4:
                     _b = options || {}, _c = _b.width, width = _c === void 0 ? 0 : _c, _d = _b.height, height = _d === void 0 ? 0 : _d, _e = _b.seconds, seconds = _e === void 0 ? 0 : _e;
                     time = moment().startOf('day').seconds(seconds).format('HH:mm:ss');
                     args = [];
@@ -145,33 +148,35 @@ var getAndCacheImageByVideo = function (uri, fileNameWithoutWxt, options) {
                                 ok(isCreated);
                             });
                         })];
-                case 4:
+                case 5:
                     isCreated = _f.sent();
                     if (!isCreated)
                         return [2 /*return*/, ''];
                     return [4 /*yield*/, getBlobLocalFile(output_path)];
-                case 5:
-                    imageData = _f.sent();
-                    return [4 /*yield*/, removeFile(output_path)];
                 case 6:
+                    imageData = _f.sent();
+                    if (!!isCachingImages) return [3 /*break*/, 8];
+                    return [4 /*yield*/, removeFile(output_path)];
+                case 7:
                     _f.sent();
-                    return [2 /*return*/, imageData];
+                    _f.label = 8;
+                case 8: return [2 /*return*/, imageData];
             }
         });
     });
 };
-var getBacketFileAndImage = function (fileName, backetName) {
+var getBacketFileAndImage = function (fileName) {
     if (fileName === void 0) { fileName = ''; }
-    if (backetName === void 0) { backetName = ''; }
-    return __awaiter(_this, void 0, void 0, function () {
-        var fullUrl, fileNameWithoutExt, image, _a, video;
+    return __awaiter(void 0, void 0, void 0, function () {
+        var fullUrl, fileNameWithoutExt, resultImageName, image, _a, video;
         return __generator(this, function (_b) {
             switch (_b.label) {
                 case 0:
                     if (!fileName)
                         return [2 /*return*/, {}];
-                    fullUrl = backetDefaultBase + "/" + backetName + "/" + fileName;
+                    fullUrl = backetDefaultBase + "/" + backetDefaultName + "/" + fileName;
                     fileNameWithoutExt = fileName.replace(/(.mp4)$/, '');
+                    resultImageName = fileNameWithoutExt + "." + OUT_FORMAT;
                     return [4 /*yield*/, getAndCacheImageByVideo(fullUrl, fileNameWithoutExt)];
                 case 1:
                     image = (_b.sent()) || '';
@@ -179,15 +184,36 @@ var getBacketFileAndImage = function (fileName, backetName) {
                 case 2:
                     _a = (_b.sent()).data, video = _a === void 0 ? {} : _a;
                     console.log(((image === null || image === void 0 ? void 0 : image.length) || 0), 'size image', ((video === null || video === void 0 ? void 0 : video.length) || 0), 'size video');
-                    return [2 /*return*/, { video: video, image: image }];
+                    return [2 /*return*/, { video: video, image: image, resultImageName: resultImageName }];
             }
         });
     });
 };
+var uploadFileToBacket = function (file, fileName) {
+    try {
+        var config = {
+            endpoint: backetDefaultBase,
+            region: 'ru-central1',
+            accessKeyId: process.env.KEY_ID,
+            secretAccessKey: process.env.KEY
+        };
+        var s3 = new AWS.S3(config);
+        var Body = Buffer.from(file, 'binary');
+        s3.putObject({
+            Body: Body,
+            Bucket: backetDefaultName,
+            Key: fileName
+        }).promise().then(function (data) {
+            console.log(data, 'success ✅ ');
+        });
+    }
+    catch (e) {
+        console.log('uploadFileToBacket', e);
+    }
+};
 var videopreview = function (event) {
-    var _a;
     return __awaiter(this, void 0, void 0, function () {
-        var fileName, backetDefaultName, backetName, image;
+        var fileName, _a, image, video, resultImageName;
         return __generator(this, function (_b) {
             switch (_b.label) {
                 case 0:
@@ -197,18 +223,34 @@ var videopreview = function (event) {
                                 statusCode: 400,
                                 body: { message: 'file name is not correct' }
                             }];
-                    backetDefaultName = 'videopreviewer';
-                    backetName = (event === null || event === void 0 ? void 0 : event.messages) ? (_a = event.messages[0].details) === null || _a === void 0 ? void 0 : _a.bucket_id : backetDefaultName;
-                    console.log(backetName, 'backetName', fileName, 'fileName');
-                    return [4 /*yield*/, getBacketFileAndImage(fileName, backetName)];
+                    return [4 /*yield*/, getBacketFileAndImage(fileName)];
                 case 1:
-                    image = (_b.sent()).image;
+                    _a = _b.sent(), image = _a.image, video = _a.video, resultImageName = _a.resultImageName;
+                    uploadFileToBacket(image, resultImageName);
                     return [2 /*return*/, {
                             statusCode: 200,
-                            body: String("data:" + image)
+                            body: { image: image, video: video }
                         }];
             }
         });
     });
 };
+var test = function () { return __awaiter(void 0, void 0, void 0, function () {
+    var testFileName, res;
+    return __generator(this, function (_a) {
+        switch (_a.label) {
+            case 0:
+                testFileName = 'cat.mp4';
+                return [4 /*yield*/, uploadFileToBacket(fs.readFileSync(testFileName, 'binary'), testFileName)];
+            case 1:
+                _a.sent();
+                return [4 /*yield*/, videopreview({ messages: [{ details: { object_id: testFileName } }] })];
+            case 2:
+                res = _a.sent();
+                console.log(res.statusCode, '= result code should be 200 ⚙️ ');
+                return [2 /*return*/];
+        }
+    });
+}); };
+test();
 module.exports.handler = videopreview;
